requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
transformers>=4.36.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
pysignalclirestapi>=0.3.20
python-dotenv>=1.0.0
pytest>=7.4.0
langdetect>=1.0.9
